+++ sdk/python/tests/conftest.py
@@ -43,7 +43,7 @@ from tests.integration.feature_repos.repo_configuration import (  # noqa: E402
 from tests.integration.feature_repos.universal.data_sources.file import (  # noqa: E402
     FileDataSourceCreator,
 )
-from tests.utils.http_server_test_utils import check_port_open, free_port  # noqa: E402
+from tests.utils.http_server_functions import check_port_open, free_port  # noqa: E402
 
 logger = logging.getLogger(__name__)
