+++ sdk/python/tests/integration/offline_store/test_universal_historical_retrieval.py
@@ -27,7 +27,7 @@ from tests.integration.feature_repos.universal.entities import (
     driver,
     location,
 )
-from tests.utils.feature_store_test_utils import (
+from tests.utils.feature_store_test_functions import (
     assert_feature_service_correctness,
     assert_feature_service_entity_mapping_correctness,
     get_expected_training_df,
