+++ sdk/python/tests/integration/registration/test_universal_cli.py
@@ -8,12 +8,12 @@ from assertpy import assertpy
 
 from feast.feature_store import FeatureStore
 from tests.integration.feature_repos.repo_configuration import Environment
-from tests.utils.cli_utils import CliRunner, get_example_repo
-from tests.utils.e2e_test_utils import (
+from tests.utils.cli_helper_functions import CliRunner, get_example_repo
+from tests.utils.e2e_test_validation_functions import (
     NULLABLE_ONLINE_STORE_CONFIGS,
     make_feature_store_yaml,
 )
-from tests.utils.online_read_write_test_utils import basic_rw_test
+from tests.utils.basic_read_write_test import basic_rw_test
 
 
 @pytest.mark.integration
