+++ sdk/python/tests/unit/cli/test_cli_chdir.py
@@ -2,7 +2,7 @@ import tempfile
 from datetime import datetime, timedelta
 from pathlib import Path
 
-from tests.utils.cli_utils import CliRunner
+from tests.utils.cli_helper_functions import CliRunner
 
 
 def test_cli_chdir() -> None:
