+++ sdk/python/tests/utils/cli_helper_functions.py
