+++ sdk/python/tests/utils/e2e_test_validation_functions.py
