+++ sdk/python/tests/utils/log_test_helper_functions.py
